pyproject.toml
[project]
name = "prolific-integration-for-ango"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
